save user theme

+++ Models/Rating.cs
@@ -9,7 +9,7 @@ namespace CollaborativeBlog.Models
     public class Rating
     {
         public int RatingId { get; set; }
-        public int RatingNumber { get; set; }
+        public double RatingNumber { get; set; }
         public int PostId { get; set; }
         public Post Post { get; set; }
         public string UserId { get; set; }
